back/requirements.txt
flask~=1.1.2
flask_login
flask_migrate
flask_wtf
flask_sqlalchemy==2.*
email_validator
python-decouple
gunicorn
Pillow~=8.0.1
netifaces~=0.10.9
opencv-python~=4.4.0.44
imutils~=0.5.3
itsdangerous~=1.1.0
wtforms~=2.3.3
sqlalchemy~=1.3.20
jinja2~=2.11.2
numpy~=1.19.3